+++ VMIRestaurant/domain/restaurant/Ingredient.cs
@@ -6,8 +6,8 @@ namespace VMIRestaurant.domain.restaurant
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Stock { get; set; }
+        public double Stock { get; set; }
         public UnitMeasure UnitMeasure { get; set; }
-        public float AmountInPortion { get; set; }
+        public double AmountInPortion { get; set; }
     }
 }
